LoanAgent and Lender Applications on individual ports
