+++ common/src/main/java/org/example/registry/ParticipantRole.java
@@ -1,0 +1,5 @@
+package org.example.registry;
+
+public enum ParticipantRole {
+    LOAN_AGENT, LENDER, DERIVED_DATA_PROVIDER, COLLECTIONS_PARTNER, KYC_PARTNER, DISBURSEMENT_PARTNER
+}
