+++ common/src/main/java/org/example/registry/RegistryService.java
@@ -1,0 +1,8 @@
+package org.example.registry;
+
+import reactor.core.publisher.Mono;
+
+public interface RegistryService {
+    Mono<ParticipantDetail> getEntity(String entityId);
+
+}
