+++ lender/src/main/resources/application.properties
@@ -1,0 +1,12 @@
+server.port=8084
+client.id=48f8617b-384b-4e4b-9022-c17424876e65
+client.secret=oIaxOAr7yVHqG4bu65txTxn8PAHcpLOX
+participant.id=3051
+loan.agent.participant.id=3051
+
+ocen.token.generation.url=https://auth.ocen.network/realms/dev/protocol/openid-connect/token
+ocen.participant.roles.url=https://dev.ocen.network/service/participant-roles/
+
+
+
+
