+++ loanagent/src/main/resources/application.properties
@@ -1,0 +1,10 @@
+server.port=8085
+loan.agent.client.id=48f8617b-384b-4e4b-9022-c17424876e65
+loan.agent.client.secret=oIaxOAr7yVHqG4bu65txTxn8PAHcpLOX
+lender.participant.id=3005
+
+client.id=f59adada-ea8d-4f8c-b511-4161fd216c49
+client.secret=LGweLUAP7OEryjiQQK4byGppswSJYIph
+
+ocen.token.generation.url=https://auth.ocen.network/realms/dev/protocol/openid-connect/token
+ocen.participant.roles.url=https://dev.ocen.network/service/participant-roles/
