+++ loanagent/src/main/resources/loan_agent_private_public_keypair_set.json
@@ -1,0 +1,18 @@
+{
+  "keys": [
+    {
+      "p": "-9vnD0FkfGcHy-7w0GhO0E_amImDXelGeBo8S8qp8oAaivkkS6-JevUNTLdOSSqws7uZxDM539P4nngNpbdWPjESOny5s0F9zrC4iBvvrav7Z8KNSJwV_DFxbjZ_Jos50FdipQF3MOiH0_jbXrgHi5sulKxWC48tLj6k8Ey11PM",
+      "kty": "RSA",
+      "q": "8D7P1xIYKi8br4HdUQgstASWSGjYtqyc-RHF0asOswyIo4W_0RLc9Kn0G06_WV07jShhYBL4EtxYon9Vzx6c2BSrrqB2BUeA5_KAPPkIJkkkVELTQf_vIADXtblxDthfLdTXojtpzUyTTytCdHTSMKK_Oy9re0urs4JyX-Hfei0",
+      "d": "i7kYTHie6q_sOAWO0cevT3mvSboKNWywE_cO2UWIu6WiQeFkxhbRmhD0kRgpGUnzddkcNFRZ87YRLBItJkUATKy36caKrEJ9if3XfYo0N3hUpLuXovUfEXcap46pATz6KP8qV6cXQLks5sn0k0W0qi4lUZPbeDm0h4cfb5VpKcy8cGK5k2bIDiLAvZhVoRiwqEmg2RSIi5dfO1OdiJYNIDDKzM7R6HYDB1hVic4f0ROsyb84UKB9nVlUutn-xNvvc-P_LIfTWF8GD8weCKk1O-wnhtEMvtQUU4jdJOPJ3f3QPYuQYLuq-WSUCHglql3xH_HczRfPtQHf4_mFaml5EQ",
+      "e": "AQAB",
+      "use": "sig",
+      "kid": "Vhq8wo1PT-o9dXisfgZxIZ3NRW5vrkEznEybY2kARJ8",
+      "qi": "Ea4DTzvxuEUNPdAcyhhQZq3jDQfNQlnHJFPGc7-Sxd63feqw-gGUXVSltugGWKP79PKzhz2YzNCSi8OMGa6BOtsFMWu-4BX225h4WpJJjmYnOuIkKu6UOGXqdV8XoUEZyssmx6Z-7ouDyxymkl_jGZB7KLVaCinwdrqDInIYhu0",
+      "dp": "N47t7J6XRR4RSTpfYwnvgOJsrUOAeCHyiqH29R0y2izKLigBirShVj9ifK7fFSLWsf9LT02hka-X45VGJW1JGBUJGV_UMyfaUdMIEfZfq-cJYX04X3EXJHffttlMuTrzOPNd7lphaaeR9v-pnL5p47EK2H8fyBMAe3gUOHIjlmM",
+      "alg": "RS512",
+      "dq": "e5Pjtz6QfYpz1EiZvifVTtgZMa_04apSqG-y8NHrosAsvOh8rZ_lVljhvvY3yVEC4-5EgHXhc1oyoYucaQjXJvMjCJVzf-2y6hy8SCDVR1RSQqGJJ6btqumkC88Sz88OznVwYfvrxXXBRHT9HwhEYk4kzd3rjKV08caE7Ls0CBU",
+      "n": "7Fv0WqpRqwQMythilZCL6-s3g9Wu4QCGfJCY_IAZYb7HvOOC7IEWl2OElF0Ncyr-0o4ibErlFIYk_dCrb6WK8m8sfLFHBq58ElAR8VeDJleYDYjMJT8k-zDOXdAwhvHXkr3uQb54PA3-LQ7IuiTfBrgJnoC6J-tBpnLWRQtHInJgK-OlyCTGwmcxELwSrHvK9ey0FZamTUf99nAMYvcuX1VtaaNH-6LycDstfhavYm1AGALSBOOHLbNJsJxBMdeuCiK9XEwzH-geGFA9XM2hN4qr9QvhFBpGp9wdq-JMKHtkMXWx-4BfH-sbnXawDSuJZo9xrUs5yWZZEGGHNh88tw"
+    }
+  ]
+}
