--- src/main/java/com/example/ocenauthentication/registry/ParticipantRole.java
@@ -1,5 +1,0 @@
-package com.example.ocenauthentication.registry;
-
-public enum ParticipantRole {
-    LOAN_AGENT, LENDER, DERIVED_DATA_PROVIDER, COLLECTIONS_PARTNER, KYC_PARTNER, DISBURSEMENT_PARTNER
-}
