--- src/main/java/com/example/ocenauthentication/registry/RegistryService.java
@@ -1,8 +1,0 @@
-package com.example.ocenauthentication.registry;
-
-import reactor.core.publisher.Mono;
-
-public interface RegistryService {
-    Mono<ParticipantDetail> getEntity(String entityId);
-
-}
