--- src/main/java/com/example/ocenauthentication/registry/Token.java
@@ -1,14 +1,0 @@
-package com.example.ocenauthentication.registry;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Token {
-    @JsonProperty("access_token")
-    private String accessToken;
-}
