--- src/main/resources/application.properties
@@ -1,7 +1,0 @@
-server.port=8083
-lender.client.id=
-lender.client.secret=
-la.participant.id=1802
-
-ocen.token.generation.url=https://auth.ocen.network/realms/dev/protocol/openid-connect/token
-ocen.participant.roles.url=https://dev.ocen.network/service/participant-roles/
