--- src/main/resources/lender_private_public_keypair_set.json
@@ -1,19 +1,0 @@
-// sample Key Pair set 
-{
-  "keys": [
-    {
-      "p": "8kyMZs3bYQkiW8bVZJ2i0jfuPdTRVUQ2nP4KflScvewE-7VUHwvm3nck8jdsZHHXxzfxf8cMWwWmU9_2Vhq0XA4K0vnh4vXM1PDN-a_51QkrQYOZf5nZElSYza77znOZ3lqiClj3f_nP2sZ7lMnxWZp5fheHZdLHo9lmpW14J_c",
-      "kty": "RSA",
-      "q": "sOmzsHxqhfJ8N4RWFEeYEoXqv-6jLq965o3y_5oMM51hecJ-NvmcZF0EVr-A3GQ0OF-xPJUEM5B4VGOIcUxuH_SL95WdJFjt428D0NwW3jJTI9fKWGJwhaHpGb3uwwOMSM6-ULlJIfBR10mhbHPq7pVwK59wqkyFU4wb1WpUmrE",
-      "d": "IIo-Eqwp7upqryzO-0_hxiWFcnoiYyguPlnM7zCCMSH28rE2t34e0neiZPyFX3H0ETyJELzUVqB00wcSQX1OsqpJxHVbcxXKDK1NT1cgJisIV7c9YqbSv7STFyW_lM3afGzbL0ObemRpmpfgnBdGesKmogfPHvICEsKJADRvvRlyG4vwWTXE3gRNT70kx5PKm6Rfa-JTXAFr4FR3fefe1UycbWq0eZLoBmNGmuO-1VaNbN9ZxqTJ_jIncevAUSJWBlm0kcJoOE3YtmlsRujK0im-kf0YRkN5TmvSiVLdXTFPNgsrJQDyS80AbLM2kx8t6FICdBIkDI09_5EFcYH4AQ",
-      "e": "AQAB",
-      "use": "sig",
-      "kid": "9584418e-76ff-4e30-b7f9-9d8bc17588c1",
-      "qi": "RiRA_7vwA_ccQxOBKYvPYXC7PGC3_di880zVOku0pRCAHpRHKwzbmAkBp030NNwTWsv-PzPPfccgg32I899P_t6bwphSij3z3CQ-kQWRismS8HFUzHgGG0BmQ2eMVy7JMJG9IVEf9mxqP0B8U16YIQ_JH5D_bAwcd-uDe3u2YDw",
-      "dp": "FK6IBXZo1V2MOz07lb0G2t6_TTPPMnICH2Bn0OS0BFoTCtEtfwDrEe4ljy8_50r4Wh7iVD_PRC21Rs-FHVBmOL1SGzQNHexRMZPkYP2uW4rO0GbFQWkGaFQkPtpC0Bp59crSrgNuDNRr84ahoc-i9D0FRcgIXJ6jP88cvd4cNX0",
-      "alg": "RS256",
-      "dq": "ZF-D6qr8SRkF7v2vVPw1xLSkNtcvUGqRYUPCTdDaAWLWjh2SPKCsYakAsDMg6moT0rRAO64UXykljEOYgyepB4pY61TAfoHSHoq68IIem1yUHS7uh-z6CGePeGJxl8Voj1p1Uhl967qMxIfrBTxTW5__r1BbvNODhZ5loFUewVE",
-      "n": "p3HSRQemfnq5wrk1jNdrN3aVtWDm3DE780wgV20NimAAaktyBjUxN8SZqSTfHWnhMJMulA7lYEHWlSTE_wbB7dlV2BpeuPMT_UDYDZjA6RGok-5qAgOlcBEZ9tndKCZB0IyyF8ZTk-__-uJQzaqBAXhPNa1esEOO48qHIWVOLoCoIuojnV7y9V8f6uXGLR71IdjkXJNf5fXaaozDxNaL2ghDLVoXtzBRfOk5fZElrpjrFAh_X-UnFq9-iVkyHTfavMtuz6LtePtiRNjrre-b-afl6hjLME1xusqbTssvYEnHeGEF2T_Qe3v91fmxnDURBXV1F92IYJzv73W3Wyo3xw"
-    }
-  ]
-}
